+++ packages/keyring-eth-ledger-bridge/src/ledger-iframe-bridge.test.ts
@@ -571,6 +571,127 @@ describe('LedgerIframeBridge', function () {
       // eslint-disable-next-line @typescript-eslint/unbound-method
       expect(bridge.iframe?.contentWindow?.postMessage).toHaveBeenCalled();
     });
+
+    describe('getAppNameAndVersion', function () {
+      it('sends and processes a successful ledger-get-app-name-and-version message', async function () {
+        const payload = {
+          appName: 'Ethereum',
+          version: '1.9.0',
+        };
+
+        stubKeyringIFramePostMessage(bridge, (message) => {
+          expect(message).toStrictEqual({
+            action: IFrameMessageAction.LedgerGetAppNameAndVersion,
+            messageId: 1,
+            target: LEDGER_IFRAME_ID,
+            params: {},
+          });
+
+          sendMessageToBridge(bridge, {
+            action: IFrameMessageAction.LedgerGetAppNameAndVersion,
+            messageId: 1,
+            success: true,
+            payload,
+          });
+        });
+
+        const result = await bridge.getAppNameAndVersion();
+
+        expect(result).toBe(payload);
+
+        // eslint-disable-next-line @typescript-eslint/unbound-method
+        expect(bridge.iframe?.contentWindow?.postMessage).toHaveBeenCalled();
+      });
+
+      it('throws an error when a ledger-get-app-name-and-version message is not successful', async function () {
+        const errorMessage = 'Ledger Error';
+
+        stubKeyringIFramePostMessage(bridge, (message) => {
+          expect(message).toStrictEqual({
+            action: IFrameMessageAction.LedgerGetAppNameAndVersion,
+            messageId: 1,
+            target: LEDGER_IFRAME_ID,
+            params: {},
+          });
+
+          sendMessageToBridge(bridge, {
+            action: IFrameMessageAction.LedgerGetAppNameAndVersion,
+            messageId: 1,
+            success: false,
+            payload: { error: new Error(errorMessage) },
+          });
+        });
+
+        await expect(bridge.getAppNameAndVersion()).rejects.toThrow(
+          errorMessage,
+        );
+
+        // eslint-disable-next-line @typescript-eslint/unbound-method
+        expect(bridge.iframe?.contentWindow?.postMessage).toHaveBeenCalled();
+      });
+    });
+
+    describe('getAppConfiguration', function () {
+      it('sends and processes a successful ledger-get-app-configuration message', async function () {
+        const payload = {
+          arbitraryDataEnabled: 1,
+          erc20ProvisioningNecessary: 0,
+          starkEnabled: 0,
+          starkv2Supported: 0,
+          version: '1.9.0',
+        };
+
+        stubKeyringIFramePostMessage(bridge, (message) => {
+          expect(message).toStrictEqual({
+            action: IFrameMessageAction.LedgerGetAppConfiguration,
+            messageId: 1,
+            target: LEDGER_IFRAME_ID,
+            params: {},
+          });
+
+          sendMessageToBridge(bridge, {
+            action: IFrameMessageAction.LedgerGetAppConfiguration,
+            messageId: 1,
+            success: true,
+            payload,
+          });
+        });
+
+        const result = await bridge.getAppConfiguration();
+
+        expect(result).toBe(payload);
+
+        // eslint-disable-next-line @typescript-eslint/unbound-method
+        expect(bridge.iframe?.contentWindow?.postMessage).toHaveBeenCalled();
+      });
+
+      it('throws an error when a ledger-get-app-configuration message is not successful', async function () {
+        const errorMessage = 'Ledger Error';
+
+        stubKeyringIFramePostMessage(bridge, (message) => {
+          expect(message).toStrictEqual({
+            action: IFrameMessageAction.LedgerGetAppConfiguration,
+            messageId: 1,
+            target: LEDGER_IFRAME_ID,
+            params: {},
+          });
+
+          sendMessageToBridge(bridge, {
+            action: IFrameMessageAction.LedgerGetAppConfiguration,
+            messageId: 1,
+            success: false,
+            payload: { error: new Error(errorMessage) },
+          });
+        });
+
+        await expect(bridge.getAppConfiguration()).rejects.toThrow(
+          errorMessage,
+        );
+
+        // eslint-disable-next-line @typescript-eslint/unbound-method
+        expect(bridge.iframe?.contentWindow?.postMessage).toHaveBeenCalled();
+      });
+    });
   });
 
   describe('setOption', function () {